+++ source/Energinet.DataHub.MarketParticipant.Integration.Model/documents/release-notes/release-notes.md
@@ -3,3 +3,12 @@
 ## Version 1.1.0
 
 - Preparing package for release.
+
+## Version 1.2.0
+
+- 
+
+## Version 1.3.0
+
+- Adding shared contract parser returning base integration event.
+- Contract related parsers marked as internal.